Moved firebase from /app to /lib (#21)
