+++ fullstack/src/app/login/page.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import { useRouter } from "next/navigation";
 import { WandSparkles } from 'lucide-react';
 import { signInWithEmailAndPassword } from 'firebase/auth';
-import { auth } from '@/app/firebase/config'
+import { auth } from '@/lib/firebase/config'
 import { useToast } from "@/hooks/use-toast"
 import { LoginData } from '@/types/authentication';
 import { LoginForm } from '@/components/LoginForm/login-form'
