+++ fullstack/src/lib/firebase/config.ts
