+++ fullstack/src/lib/utils.ts
@@ -1,3 +1,5 @@
+// This whole file is from shadcn iirc
+
 import { clsx, type ClassValue } from "clsx"
 import { twMerge } from "tailwind-merge"
